Added a missing client update when the mute status is manually cancelled by a GM (followup to r12079)
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@12169 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/socket.c
@@ -575,7 +575,7 @@ int realloc_writefifo(int fd, size_t addition)
 		while( session[fd]->wdata_size + addition > newsize ) newsize += newsize;
 	}
 	else
-	if( session[fd]->max_wdata >= 2*(session[fd]->flag.server?FIFOSIZE_SERVERLINK:WFIFO_SIZE)
+	if( session[fd]->max_wdata >= (size_t)2*(session[fd]->flag.server?FIFOSIZE_SERVERLINK:WFIFO_SIZE)
 		&& (session[fd]->wdata_size+addition)*4 < session[fd]->max_wdata )
 	{	// shrink rule, shrink by 2 when only a quarter of the fifo is used, don't shrink below nominal size.
 		newsize = session[fd]->max_wdata / 2;
